chore: reduce duplication
